filter points in boxes that doesn't hit the car

+++ src/neural_car_detector/dbscan.h
@@ -7,9 +7,8 @@ namespace dbscan
 {
 
 struct Point {
-	int x;
-
-	int y;
+	double x;
+	double y;
 };
 
 typedef std::vector<Point> Cluster;
